+++ src/browser-lib/fixtures/blogListItems.ts
@@ -1,0 +1,54 @@
+import { BlogListProps } from "../../components/BlogList/BlogList";
+
+const blogListItems: BlogListProps["items"] = [
+  {
+    titleTag: "h3",
+    title: "Webinar about school",
+    snippet:
+      "Preview, plan and customise each element of your work to meet your needs - in and out the classroom.",
+    href: "/",
+    contentType: "webinar",
+  },
+  {
+    titleTag: "h3",
+    title: "Blog about learning",
+    snippet:
+      "Preview, plan and customise each element of your work to meet your needs - in and out the classroom.",
+    href: "/",
+    contentType: "blog-post",
+  },
+  {
+    titleTag: "h3",
+    title: "Webinar from a maths teacher",
+    snippet:
+      "Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+    href: "/",
+    contentType: "webinar",
+  },
+  {
+    titleTag: "h3",
+    title: "Webinar about making lunch break all day",
+    snippet:
+      "Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.",
+    href: "/",
+    contentType: "webinar",
+  },
+  {
+    titleTag: "h3",
+    title: "Webinar about making rest in the evening",
+    snippet:
+      "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
+    href: "/",
+    contentType: "webinar",
+  },
+  {
+    titleTag: "h3",
+    title: "Blog post about hot to have fun at work",
+    snippet:
+      "Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
+    href: "/",
+    contentType: "blog-post",
+  },
+];
+
+export default blogListItems;