+++ src/components/Card/Card.tsx
@@ -1,21 +1,13 @@
 import styled from "styled-components";
 
-import background, { BackgroundProps } from "../../styles/utils/background";
-import position from "../../styles/utils/position";
-import spacing, { SpacingProps } from "../../styles/utils/spacing";
-import { BoxProps } from "../Box";
+import Flex from "../Flex/Flex";
 
 /**
  * @todo ts should be able to infer card props from e.g. 'spacing'
  * and 'background' util types
  */
-export type CardProps = SpacingProps & BackgroundProps & BoxProps;
-const Card = styled.div<CardProps>`
-  display: flex;
-  flex-direction: column;
-  ${spacing}
-  ${background}
-  ${position}
+
+const Card = styled(Flex)`
   min-height: 200px;
   border-radius: 12px;
   flex-grow: 1;
@@ -23,7 +15,7 @@ const Card = styled.div<CardProps>`
 
 Card.defaultProps = {
   pa: 24,
-  // background: "inYourFace",
+  flexDirection: "column",
 };
 
 export default Card;